pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm[simple]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "gmol-core"
description = "Core components of GalaxyMol"
readme = "README.md"
authors = [{ name = "Seoklab members", email = "contact@seoklab.org" }]
maintainers = [{ name = "Nuri Jung", email = "jnooree@snu.ac.kr" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "numpy",
    "nurikit>=0.1.0a16,<0.2.0",
    "pandas>=2.0,<3.0",
    "pydantic>=2.7,<3.0",
    "rdkit>=2024.3",
    "scipy>=1.15.3,<2.0.0",
    "tqdm",
]
dynamic = ["version"]

[project.urls]
repository = "https://github.com/seoklab/gmol-core.git"

[dependency-groups]
dev = [
    "debugpy",
    "jupyter",
    "pre-commit>=4.5.1,<5.0.0",
    { include-group = "lint" },
    { include-group = "test" },
    { include-group = "typecheck" },
]
lint = ["ruff"]
test = ["coverage", "pytest"]
typecheck = ["mypy"]
docs = ["pydata-sphinx-theme", "sphinx>=8.2,<9.0"]

[tool.uv]
add-bounds = "major"
environments = ["sys_platform == 'linux'"]
constraint-dependencies = [
    "pygls<2.0", # vscode requires pygls<2.0
]

[tool.uv.dependency-groups]
docs = { requires-python = ">=3.11" }

[tool.ruff]
line-length = 79
src = ["src"]
extend-include = ["*.pyi", "*.ipynb"]
extend-exclude = [".github/**", "*.pyc"]

[tool.ruff.lint]
extend-select = [
    "W6",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "INP",
    "T20",
    "I",
    "N80",
    "S102",
    "S108",
    "S306",
    "S307",
    "S60",
    "B",
    "A",
    "PT",
    "PTH",
    "D301",
    "D418",
    "PLE",
    "PLW",
    "NPY",
    "UP",
    "RUF",
]
ignore = [
    "E74",
    "F403",
    "F405",
    "N803",
    "N806",
    "S603",
    "S607",
    "B008",
    "B905",
    "PTH123",
    "PLW2901",
    "NPY002",
    "UP034",
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
    "F401", # Imported but unused
    "F403", # Wildcard imports
]

[tool.ruff.lint.isort]
combine-as-imports = true
no-lines-before = ["local-folder"]

[tool.ruff.lint.flake8-import-conventions]
banned-from = [
    "numpy",
    "pandas",
    "scipy.spatial.distance",
    "scipy.cluster.hierarchy",
    "subprocess",
]

[tool.ruff.lint.flake8-import-conventions.extend-aliases]
networkx = "nx"
"numpy.linalg" = "LA"
"scipy.spatial.distance" = "D"
"scipy.cluster.hierarchy" = "hc"
"subprocess" = "sp"

[tool.mypy]
files = ["src/"]
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"

plugins = ["pydantic.mypy"]

check_untyped_defs = true
disallow_any_explicit = true
disallow_any_generics = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true

allow_redefinition_new = true
local_partial_types = true
disable_error_code = ["import-untyped", "no-redef"]

show_error_context = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = ["rdkit.*", "pandas.*"]

follow_imports = "skip"
follow_imports_for_stubs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.pytest]
testpaths = ["tests"]

[tool.coverage.run]
command_line = "-m pytest -v -s"
branch = true
relative_files = true
dynamic_context = "test_function"
source = ["src"]

[tool.coverage.report]
skip_covered = true
skip_empty = true
sort = "-cover"
exclude_also = [
    "def __repr__",
    "@abstractmethod",
    "if __name__ == ['\"]__main__['\"]:",
]

[tool.coverage.html]
show_contexts = true
skip_covered = false
skip_empty = true

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.10"
pythonPlatform = "Linux"

reportGeneralTypeIssues = "information"
reportUnusedImport = "warning"
reportUnusedClass = "warning"
reportUnusedFunction = "warning"
reportUnusedVariable = "warning"
reportDuplicateImport = "warning"
reportWildcardImportFromLibrary = "none"
reportOptionalSubscript = "warning"
reportOptionalMemberAccess = "warning"
reportOptionalCall = "warning"
reportOptionalIterable = "warning"
reportOptionalContextManager = "warning"
reportOptionalOperand = "warning"
reportPrivateImportUsage = "none"
reportIncompatibleMethodOverride = "none"
reportIncompatibleVariableOverride = "warning"
reportInconsistentConstructor = "warning"
reportOverlappingOverload = "warning"
reportUninitializedInstanceVariable = "none"
reportInvalidStringEscapeSequence = "error"
reportUnnecessaryIsInstance = "warning"
reportUnnecessaryCast = "warning"
reportUnnecessaryComparison = "warning"
reportUnnecessaryContains = "warning"
reportAssertAlwaysTrue = "warning"
reportSelfClsParameterName = "warning"
reportImplicitStringConcatenation = "information"
reportInvalidStubStatement = "warning"
reportIncompleteStub = "warning"
reportUnnecessaryTypeIgnoreComment = "none"
reportMatchNotExhaustive = "warning"
reportShadowedImports = "warning"
reportPossiblyUnboundVariable = "none"
reportMissingModuleSource = "none"

# Covered by mypy
reportAssignmentType = "none"
reportAttributeAccessIssue = "none"
reportArgumentType = "none"
reportCallIssue = "none"
reportReturnType = "none"
reportAny = "none"
reportRedeclaration = "none"
reportImplicitOverride = "none"
reportUnusedCallResult = "none"
reportMissingTypeArgument = "none"
reportMissingParameterType = "none"
reportUnknownArgumentType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
reportUnknownLambdaType = "none"
